requirements.txt
psycopg2-binary
FastAPI
uvicorn